Conflict in test_DSA_sparse_score_recompute.py: both sides appended tests (this branch's Rule 8 detectors,
#1225's ring-phase and SMEM-budget regressions); kept both. #1225's 227 KiB budget lives in _interface_sm100,
which this branch's score-recompute plans already plan through.

+++ python/cudnn/sdpa/fwd/kernels/sm100/prefill_d512_f16.py
@@ -2158,7 +2158,7 @@ def _host(
             thd_lens_form,
             cutlass.Int32(QH // HEADS_PER_TILE),
             cutlass.Int32(B),
-            cutlass.Int32(o_tensor.stride[1]),
+            cutlass.Int64(o_tensor.stride[1]),
             cutlass.Int32(CFG.TILES_Q * CFG.TILE_M * CFG.CTA_MMA),
             n_thd_units,  # persistent cluster count; also seeds the claim counter
         ).launch(grid=(1, 1, 1), block=(THD_SETUP_THREADS, 1, 1), stream=stream)
